feat(platform): complete P1 - StateFs migration for tools/cli + Linux platform
Migrate files/file_write/cli to StateFs abstraction and add Linux platform support.

Changes:
- Migrate tools/files.rs: remove cfg dual-path, use injected StateFs
- Migrate tools/file_write.rs: write/append via StateFs, protected paths
- Migrate cli session_list: use platform.state_fs().list_dir()
- Add LinuxStateFs: std::fs + mutex, auto create parent dirs
- Add LinuxPlatform: mirrors Esp32Platform structure
- Add state_root.rs: unified mount path (ESP /spiffs, Linux env/default)
- Add normalize_state_rel_path: shared path validation for tools/platform
- Refactor spiffs/wifi/http_client into submodules

Verification:
- ESP + Linux musl targets both compile successfully
- tools/cli no longer import platform::spiffs directly
- Net -925 lines (removed dual-path implementations)

Refs: dev-docs/platform-isolation-plan.md §7.1, §11 P1

+++ src/platform/http_server/handlers/system_info.rs
@@ -54,10 +54,10 @@ fn days_to_ymd(days: u64) -> (u32, u32, u32) {
     }
     let (mo, day) = if m >= 12 {
         let dec_start = acc - mon_days[11];
-        (12u32, ((d - dec_start + 1) as u32).min(mon_days[11]))
+        (12u32, ((d - dec_start + 1) as u32).min(mon_days[11] as u32))
     } else {
         let day_raw = (d - acc + 1) as u32;
-        (m as u32 + 1, day_raw.min(mon_days[m]))
+        (m as u32 + 1, day_raw.min(mon_days[m] as u32))
     };
     (y, mo, day)
 }
